requirements.txt
python-dotenv
langchain
openai
google-api-python-client # Google API
google-search-results # SerpAPI
youtube-transcript-api
chromadb
tiktoken
faiss-cpu
wikipedia
cohere
streamlit_chat
chainlit
git+https://github.com/openai/whisper.git
judini
pandas
pillow